breaking functionality of syntheaTo09 out into a set of utility functions to support use in other mappers.  Move existing utility functions into a separate common.js file so it could be used within the mcode09 utils without circular references.  Added a mapper for Cerner TNM, Staging and ER,PR,HER2 receptors based off of example data.  Added another value to the set of primaryCancerType codes for example test patient
