+++ src/mapping/mappers/index.js
@@ -1,7 +1,8 @@
 const syntheaToV05 = require('./syntheaToV05');
 const syntheaToV09 = require('./syntheaToV09');
-
+const cerner = require('./cerner');
 module.exports = {
     syntheaToV05,
-    syntheaToV09
+    syntheaToV09,
+    cerner
 };
