+++ src/mapping/mappers/syntheaToV09.js
@@ -1,26 +1,7 @@
 const { buildMappers } = require('../mapper');
-const { applyProfile, addExtension } = require('../../utils');
-const fhirpath = require('fhirpath');
-const _ = require('lodash');
+const { applyProfile, applyProfileFunction } = require('../../utils/common');
+const mcodeUtils09 = require('../../utils/mcodeUtils09');
 
-const applyProfileFunction = (profile) => {
-    // return an anonymous function wrapper to apply this specific profile to given resources
-    return (resource, _context) => applyProfile(resource, profile);
-};
-
-// FHIRPath helper. FHIRPath tends to return things that are JS truthy (like empty arrays)
-// when we would expect a null or other falsy value instead
-// TODO: reference the same function here and in mapper
-const isTrue = (arg) => {
-  if (Array.isArray(arg) ){
-    return arg.find(i => isTrue(i));
-  } else if (typeof arg === 'object'){
-    return !_.isEmpty(arg);
-  } else if (typeof arg === 'string' && arg === 'false'){
-    return false;
-  }
-  return arg;
-};
 
 const defaultProfile = (resourceType) => {
     switch (resourceType) {
@@ -47,129 +28,6 @@ const defaultProfile = (resourceType) => {
     }
 };
 
-// codes are listed with the modules they are used in
-// (B) = Breast Cancer, (C) = Colorectal, (L) = Lung, (P) = Prostate
-const fhirPathVariables = {
-    primaryCancerConditionCodes: [
-        '254837009', // Malignant neoplasm of breast (disorder) (B)
-        '93761005', // Primary malignant neoplasm of colon (C)
-        '109838007', // Overlapping malignant neoplasm of colon (C)
-        '363406005', // Malignant tumor of colon (C)
-        '94260004', // Secondary malignant neoplasm of colon (C) -- note this is a "secondary" code but is intended to be a primary cancer
-        '254637007', // Non-small cell lung cancer (disorder) (L)
-        '254632001', // Small cell carcinoma of lung (disorder) (L)
-        '424132000', // Non-small cell carcinoma of lung, TNM stage 1 (disorder) (L)
-        '425048006', // Non-small cell carcinoma of lung, TNM stage 2 (disorder) (L)
-        '422968005', // Non-small cell carcinoma of lung, TNM stage 3 (disorder) (L)
-        '423121009', // Non-small cell carcinoma of lung, TNM stage 4 (disorder) (L)
-        '67811000119102', // Primary small cell malignant neoplasm of lung, TNM stage 1 (disorder) (L)
-        '67821000119109', // Primary small cell malignant neoplasm of lung, TNM stage 2 (disorder) (L)
-        '67831000119107', // Primary small cell malignant neoplasm of lung, TNM stage 3 (disorder) (L)
-        '67841000119103', // Primary small cell malignant neoplasm of lung, TNM stage 4 (disorder) (L)
-        // note that none of the 3 prostate cancer codes are in the recommended VS
-        '126906006', // Neoplasm of prostate (P)
-        '92691004', // Carcinoma in situ of prostate (disorder) (P)
-        '314994000', // Metastasis from malignant tumor of prostate (disorder) (P) -- also a "secondary" code but intended to be a primary cancer
-    ],
-
-    surgeryCodes: [
-        '396487001', // Sentinel lymph node biopsy (procedure) (B)
-        '443497002', // Excision of sentinel lymph node (procedure) (B)
-        '234262008', // Excision of axillary lymph node (procedure) (B)
-        '392021009', // Lumpectomy of breast (procedure) (B)
-        '69031006', // Excision of breast tissue (procedure) (B)
-        '387607004', // Construction of diverting colostomy (B)
-        '43075005', // Partial resection of colon (B)
-        '90470006', // Prostatectomy (P)
-    ],
-
-    radiationCodes: [
-        '447759004', // Brachytherapy of breast (procedure) (B)
-        '384692006', // Intracavitary brachytherapy (procedure) (B)
-        '113120007', // Interstitial brachytherapy (procedure) (B)
-        '33195004', // Teleradiotherapy procedure (procedure) (B)
-        '385798007', // Radiation therapy care (regime/therapy) (B)
-        '108290001', // Radiation oncology AND/OR radiotherapy (procedure) (B)
-    ],
-
-    tumorMarkerTestCodes: [
-        '85319-2', // HER2 [Presence] in Breast cancer specimen by Immune stain (B)
-        '85318-4', // HER2 [Presence] in Breast cancer specimen by FISH (B)
-        '85337-4', // Estrogen receptor Ag [Presence] in Breast cancer specimen by Immune stain (B)
-        '85339-0', // Progesterone receptor Ag [Presence] in Breast cancer specimen by Immune stain (B)
-        '2857-1', // Prostate specific Ag [Mass/volume] in Serum or Plasma (P)
-    ]
-};
-
-const PRIMARY_CANCER_CONDITION_FILTER = 'Condition.code.coding.where($this.code in %primaryCancerConditionCodes)';
-const PRIMARY_CANCER_CONDITION_PATH = fhirpath.compile(PRIMARY_CANCER_CONDITION_FILTER, fhirPathVariables);
-
-// specific instance of the "find" function with a precompiled fhirpath, possibly a premature optimization but things are already slow on IE
-const findPrimaryCancerCondition = (context, resource = null) => {
-    if (typeof context === 'object' && context.entry) {
-        // extract the entries from the bundle
-        context = context.entry.map(e => e.resource);
-        // otherwise assume it's an array of resources anyway
-    }
-
-    const primaryCancers = context.filter(  r => isTrue( PRIMARY_CANCER_CONDITION_PATH(r) )  );
-
-    if (primaryCancers.length === 0) {
-        return null;
-    }
-
-    // TODO: use timestamps to figure out which one, if there are multiple
-    return primaryCancers[0];
-};
-
-const find = (context, path, options={}) => {
-     if (typeof context === 'object' && context.entry) {
-        // extract the entries from the bundle
-        context = context.entry.map(e => e.resource);
-        // otherwise assume it's an array of resources anyway
-    }
-
-    if (typeof path === 'string') {
-        path = fhirpath.compile(path);
-    }
-
-    const results = context.filter(  r => isTrue( path(r) )  );
-    if (results.length === 0) {
-        return null;
-    }
-
-    // TODO: use options to figure out which one, if there are multiple
-    return results[0];
-};
-
-const addCancerReferenceExtension = (resource, context, url) => {
-    const primaryCancer = findPrimaryCancerCondition(context, resource);
-    if (primaryCancer) {
-        const extension = {
-            url: url,
-            valueReference: {
-                reference: `Condition/${primaryCancer.id}`
-            }
-        };
-
-        addExtension(resource, extension);
-    }
-};
-
-const addRelatedCancerConditionExtension = (resource, context) => 
-    addCancerReferenceExtension(resource, context, 'http://hl7.org/fhir/us/shr/DSTU2/StructureDefinition/onco-core-RelatedCancerCondition-extension');
-
-const addCancerReasonReferenceExtension = (resource, context) =>
-    addCancerReferenceExtension(resource, context, 'http://hl7.org/fhir/us/shr/DSTU2/StructureDefinition/onco-core-CancerReasonReference-extension');
-
-
-const addStageGroupRelated = (resource, context, path) => {
-    const relatedItem = find(context, path);
-
-    if (relatedItem) {
-        resource.related.push({ target: { reference: `Observation/${relatedItem.id}` } });
-    }
-};
 
 const nthWord = (string, index) => {
     return string.split(' ')[index];
@@ -181,7 +39,7 @@ const stripParens = (string) => {
 
     const endIndex = string.lastIndexOf('(');
 
-    if (endIndex === -1) return string;
+    if (endIndex === -1) {return string;}
 
     return string.slice(0, endIndex - 1); // endIndex - 1 because there's an extra space at the end too
 };
@@ -207,7 +65,7 @@ const resourceMapping = {
             exec: (resource, context) => {
                 applyProfile(resource, 'http://hl7.org/fhir/us/shr/StructureDefinition/onco-core-CancerDiseaseStatus');
 
-                addRelatedCancerConditionExtension(resource, context);
+                mcodeUtils09.addRelatedCancerConditionExtension(resource, context);
 
                 resource.valueCodeableConcept.text = resource.valueCodeableConcept.coding[0].display = stripParens(resource.valueCodeableConcept.coding[0].display);
 
@@ -219,23 +77,23 @@ const resourceMapping = {
             exec: (resource, context) => {
                 applyProfile(resource, 'http://hl7.org/fhir/us/shr/StructureDefinition/onco-core-CancerRelatedRadiationProcedure');
 
-                addCancerReasonReferenceExtension(resource, context);
+                mcodeUtils09.addCancerReasonReferenceExtension(resource, context);
 
                 return resource;
             }
         },
         {
-            filter: 'Procedure.code.coding.where($this.code in %surgeryCodes)',
+            filter: 'Procedure.code.coding.where($this.code in %surgeryCodes))',
             exec: (resource, context) => {
                 applyProfile(resource, 'http://hl7.org/fhir/us/shr/StructureDefinition/onco-core-CancerRelatedSurgicalProcedure');
 
-                addCancerReasonReferenceExtension(resource, context);
+                mcodeUtils09.addCancerReasonReferenceExtension(resource, context);
 
                 return resource;
             }
         },
         {
-            filter: PRIMARY_CANCER_CONDITION_FILTER,
+            filter: mcodeUtils09.PRIMARY_CANCER_CONDITION_FILTER,
             exec: applyProfileFunction('http://hl7.org/fhir/us/shr/StructureDefinition/onco-core-PrimaryCancerCondition')
         },
         // { // All cancers in Synthea are intended to be primary, even if secondary codes are used
@@ -246,7 +104,7 @@ const resourceMapping = {
             filter: "Observation.code.coding.where($this.code = '21907-1')",
             exec: (resource, context) => {
                 applyProfile(resource, 'http://hl7.org/fhir/us/shr/StructureDefinition/onco-core-TNMClinicalDistantMetastasesCategory');
-                addRelatedCancerConditionExtension(resource, context);
+                mcodeUtils09.addRelatedCancerConditionExtension(resource, context);
 
                 // keep only the first word of the code. ex "T1 category (finding)" -> "T1"
                 resource.valueCodeableConcept.text = resource.valueCodeableConcept.coding[0].display = nthWord(resource.valueCodeableConcept.coding[0].display, 0);
@@ -258,7 +116,7 @@ const resourceMapping = {
             filter: "Observation.code.coding.where($this.code = '21905-5')",
             exec: (resource, context) => {
                 applyProfile(resource, 'http://hl7.org/fhir/us/shr/StructureDefinition/onco-core-TNMClinicalPrimaryTumorCategory');
-                addRelatedCancerConditionExtension(resource, context);
+                mcodeUtils09.addRelatedCancerConditionExtension(resource, context);
 
                 // keep only the first word of the code. ex "T1 category (finding)" -> "T1"
                 resource.valueCodeableConcept.text = resource.valueCodeableConcept.coding[0].display = nthWord(resource.valueCodeableConcept.coding[0].display, 0);
@@ -270,7 +128,7 @@ const resourceMapping = {
             filter: "Observation.code.coding.where($this.code = '21906-3')",
             exec: (resource, context) => {
                 applyProfile(resource, 'http://hl7.org/fhir/us/shr/StructureDefinition/onco-core-TNMClinicalRegionalNodesCategory');
-                addRelatedCancerConditionExtension(resource, context);
+                mcodeUtils09.addRelatedCancerConditionExtension(resource, context);
 
                 // keep only the first word of the code. ex "T1 category (finding)" -> "T1"
                 resource.valueCodeableConcept.text = resource.valueCodeableConcept.coding[0].display = nthWord(resource.valueCodeableConcept.coding[0].display, 0);
@@ -282,7 +140,7 @@ const resourceMapping = {
             filter: "Observation.code.coding.where($this.code = '21908-9')",
             exec: (resource, context) => {
                 applyProfile(resource, 'http://hl7.org/fhir/us/shr/StructureDefinition/onco-core-TNMClinicalStageGroup');
-                addRelatedCancerConditionExtension(resource, context);
+                mcodeUtils09.addRelatedCancerConditionExtension(resource, context);
 
                 // keep only the second word of the code. ex "Stage 1A (qualifier value)" -> "1A"
                 resource.valueCodeableConcept.text = resource.valueCodeableConcept.coding[0].display = nthWord(resource.valueCodeableConcept.coding[0].display, 1);
@@ -290,9 +148,9 @@ const resourceMapping = {
                 // find the 3 components and add them to related
                 resource.related = resource.related || [];
 
-                addStageGroupRelated(resource, context, "Observation.code.coding.where($this.code = '21905-5')"); // primary tumor
-                addStageGroupRelated(resource, context, "Observation.code.coding.where($this.code = '21906-3')"); // regional nodes
-                addStageGroupRelated(resource, context, "Observation.code.coding.where($this.code = '21907-1')"); // distant metastases
+                mcodeUtils09.addStageGroupRelated(resource, context, "Observation.code.coding.where($this.code = '21905-5')"); // primary tumor
+                mcodeUtils09.addStageGroupRelated(resource, context, "Observation.code.coding.where($this.code = '21906-3')"); // regional nodes
+                mcodeUtils09.addStageGroupRelated(resource, context, "Observation.code.coding.where($this.code = '21907-1')"); // distant metastases
 
                 return resource;
             }
@@ -301,7 +159,7 @@ const resourceMapping = {
             filter: "Observation.code.coding.where($this.code = '21901-4')",
             exec: (resource, context) => {
                 applyProfile(resource, 'http://hl7.org/fhir/us/shr/StructureDefinition/onco-core-TNMPathologicDistantMetastasesCategory');
-                addRelatedCancerConditionExtension(resource, context);
+                mcodeUtils09.addRelatedCancerConditionExtension(resource, context);
 
                 // keep only the first word of the code. ex "T1 category (finding)" -> "T1"
                 resource.valueCodeableConcept.text = resource.valueCodeableConcept.coding[0].display = nthWord(resource.valueCodeableConcept.coding[0].display, 0);
@@ -313,7 +171,7 @@ const resourceMapping = {
             filter: "Observation.code.coding.where($this.code = '21899-0')",
             exec: (resource, context) => {
                 applyProfile(resource, 'http://hl7.org/fhir/us/shr/StructureDefinition/onco-core-TNMPathologicPrimaryTumorCategory');
-                addRelatedCancerConditionExtension(resource, context);
+                mcodeUtils09.addRelatedCancerConditionExtension(resource, context);
 
                 // keep only the first word of the code. ex "T1 category (finding)" -> "T1"
                 resource.valueCodeableConcept.text = resource.valueCodeableConcept.coding[0].display = nthWord(resource.valueCodeableConcept.coding[0].display, 0);
@@ -325,7 +183,7 @@ const resourceMapping = {
             filter: "Observation.code.coding.where($this.code = '21900-6')",
             exec: (resource, context) => {
                 applyProfile(resource, 'http://hl7.org/fhir/us/shr/StructureDefinition/onco-core-TNMPathologicRegionalNodesCategory');
-                addRelatedCancerConditionExtension(resource, context);
+                mcodeUtils09.addRelatedCancerConditionExtension(resource, context);
 
                 // keep only the first word of the code. ex "T1 category (finding)" -> "T1"
                 resource.valueCodeableConcept.text = resource.valueCodeableConcept.coding[0].display = nthWord(resource.valueCodeableConcept.coding[0].display, 0);
@@ -337,7 +195,7 @@ const resourceMapping = {
             filter: "Observation.code.coding.where($this.code = '21902-2')",
             exec: (resource, context) => {
                 applyProfile(resource, 'http://hl7.org/fhir/us/shr/StructureDefinition/onco-core-TNMPathologicStageGroup');
-                addRelatedCancerConditionExtension(resource, context);
+                mcodeUtils09.addRelatedCancerConditionExtension(resource, context);
 
                 // keep only the second word of the code. ex "Stage 1A (qualifier value)" -> "1A"
                 resource.valueCodeableConcept.text = resource.valueCodeableConcept.coding[0].display = nthWord(resource.valueCodeableConcept.coding[0].display, 1);
@@ -346,9 +204,9 @@ const resourceMapping = {
                 // find the 3 components and add them to related
                 resource.related = resource.related || [];
 
-                addStageGroupRelated(resource, context, "Observation.code.coding.where($this.code = '21899-0')"); // primary tumor
-                addStageGroupRelated(resource, context, "Observation.code.coding.where($this.code = '21900-6')"); // regional nodes
-                addStageGroupRelated(resource, context, "Observation.code.coding.where($this.code = '21901-4')"); // distant metastases
+                mcodeUtils09.addStageGroupRelated(resource, context, "Observation.code.coding.where($this.code = '21899-0')"); // primary tumor
+                mcodeUtils09.addStageGroupRelated(resource, context, "Observation.code.coding.where($this.code = '21900-6')"); // regional nodes
+                mcodeUtils09.addStageGroupRelated(resource, context, "Observation.code.coding.where($this.code = '21901-4')"); // distant metastases
 
                 return resource;
             }
@@ -372,7 +230,7 @@ const resourceMapping = {
                         break;
                     case '85339-0': // PR
                         resource.code.text = 'Progesterone Receptor';
-                        resource.code.coding.unshift({ system: 'LOINC', code: '16113-3', display: 'Progesterone receptor [Interpretation] in Tissue' })
+                        resource.code.coding.unshift({ system: 'LOINC', code: '16113-3', display: 'Progesterone receptor [Interpretation] in Tissue' });
                         break;
                 }
 
@@ -388,4 +246,4 @@ const resourceMapping = {
 };
 
 
-module.exports = buildMappers(resourceMapping, fhirPathVariables);
+module.exports = buildMappers(resourceMapping, mcodeUtils09.fhirPathVariables);
